requirements.txt
PyMuPDF==1.23.6
Pillow==10.0.0 